Refactor: entire detailed screen refactored
