+++ lib/screens/stepper_page.dart
@@ -1,5 +1,5 @@
-import 'package:book_worm/screens/library.dart';
 import 'package:book_worm/services/isar_service.dart';
+import 'package:book_worm/utility/string.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_rating_bar/flutter_rating_bar.dart';
 import 'package:book_worm/models/book.dart';
@@ -9,7 +9,7 @@ import 'package:book_worm/models/user_book_entry.dart';
 class StepperPage extends StatefulWidget {
   final UserBookEntry bookEntry;
 
-  const StepperPage({Key? key, required this.bookEntry}) : super(key: key);
+  const StepperPage({super.key, required this.bookEntry});
 
   @override
   State<StepperPage> createState() => _StepperPageState();
