+++ lib/utility/string.dart
@@ -1,0 +1,6 @@
+// Add this extension method to capitalize the first letter of a string
+extension StringExtension on String {
+  String capitalize() {
+    return "${this[0].toUpperCase()}${substring(1).toLowerCase()}";
+  }
+}
