+++ lib/widgets/book_details/book_detail_summary.dart
@@ -1,0 +1,109 @@
+import 'package:book_worm/states/book_detail_state.dart';
+import 'package:flutter/material.dart';
+import 'package:provider/provider.dart';
+
+class BookDetailSummary extends StatelessWidget {
+  const BookDetailSummary({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Consumer<BookDetailState>(
+      builder: (context, state, child) {
+        return Padding(
+          padding: const EdgeInsets.symmetric(horizontal: 16.0),
+          child: Column(
+            crossAxisAlignment: CrossAxisAlignment.start,
+            children: [
+              Row(
+                children: [
+                  const Text(
+                    "Summary",
+                    style: TextStyle(
+                      fontWeight: FontWeight.w500,
+                      fontSize: 20.0,
+                    ),
+                  ),
+                  if (state.isEditMode)
+                    IconButton(
+                      onPressed: () => _openSummaryDialog(context, state),
+                      icon: const Icon(
+                        Icons.edit,
+                        size: 20,
+                      ),
+                    )
+                ],
+              ),
+              const SizedBox(height: 8.0),
+              Text(state.book.summary ?? ""),
+              const SizedBox(height: 8.0),
+            ],
+          ),
+        );
+      },
+    );
+  }
+
+  void _openSummaryDialog(BuildContext context, BookDetailState state) {
+    final TextEditingController summaryController =
+        TextEditingController(text: state.book.summary);
+
+    showDialog(
+      context: context,
+      builder: (BuildContext context) {
+        return AlertDialog(
+          title: const Text('Edit Summary',
+              style: TextStyle(fontWeight: FontWeight.bold)),
+          content: SingleChildScrollView(
+            child: Column(
+              mainAxisSize: MainAxisSize.min,
+              children: [
+                TextField(
+                  controller: summaryController,
+                  maxLines: 5,
+                  textInputAction: TextInputAction.newline,
+                  textCapitalization: TextCapitalization.sentences,
+                  decoration: InputDecoration(
+                    hintText: 'Enter book summary...',
+                    border: OutlineInputBorder(
+                      borderRadius: BorderRadius.circular(8),
+                    ),
+                    focusedBorder: OutlineInputBorder(
+                      borderRadius: BorderRadius.circular(8),
+                      borderSide: BorderSide(
+                          color: Theme.of(context).primaryColor, width: 2),
+                    ),
+                  ),
+                ),
+              ],
+            ),
+          ),
+          actions: [
+            TextButton(
+              child: const Text('Cancel'),
+              onPressed: () => Navigator.of(context).pop(),
+            ),
+            ElevatedButton(
+              onPressed: () {
+                state.updateBookSummary(summaryController.text);
+                Navigator.of(context).pop();
+              },
+              style: ElevatedButton.styleFrom(
+                backgroundColor: Theme.of(context).primaryColor,
+                shape: RoundedRectangleBorder(
+                  borderRadius: BorderRadius.circular(8),
+                ),
+              ),
+              child: const Text(
+                'Save',
+                style: TextStyle(color: Colors.white),
+              ),
+            ),
+          ],
+          shape: RoundedRectangleBorder(
+            borderRadius: BorderRadius.circular(16),
+          ),
+        );
+      },
+    );
+  }
+}
