+++ lib/widgets/image_dialog.dart
@@ -1,0 +1,24 @@
+import 'dart:io';
+
+import 'package:flutter/material.dart';
+
+class ImageDialog extends StatelessWidget {
+  final String imagePath;
+
+  const ImageDialog({super.key, required this.imagePath});
+
+  @override
+  Widget build(BuildContext context) {
+    return Dialog(
+      backgroundColor: Colors.transparent,
+      insetPadding: EdgeInsets.zero,
+      child: GestureDetector(
+        onTap: () => Navigator.of(context).pop(),
+        child: Image.file(
+          File(imagePath),
+          fit: BoxFit.contain,
+        ),
+      ),
+    );
+  }
+}
