+++ pubspec.yaml
@@ -42,7 +42,8 @@ dependencies:
   intl: ^0.19.0
   flutter_rating_bar: ^4.0.1
   smooth_page_indicator: ^1.0.0+2
-
+  provider: ^6.1.2
+  
 dev_dependencies:
   flutter_test:
     sdk: flutter
